+++ include/nstd/core.h
@@ -8,6 +8,7 @@
 #include "core/math.h"
 #include "core/platform.h"
 #include "core/pointer.h"
+#include "core/range.h"
 #include "core/slice.h"
 #include "core/str.h"
 #endif