+++ include/nstd/core/def.h
@@ -71,23 +71,6 @@ typedef enum
     NSTD_BOOL_TRUE
 } NSTDBool;
 
-/// Represents a signed range.
-typedef struct
-{
-    /// Start of the range (included).
-    NSTDInt64 start;
-    /// End of the range (exluded).
-    NSTDInt64 end;
-} NSTDIRange;
-/// Represents an unsigned range.
-typedef struct
-{
-    /// Start of the range (included).
-    NSTDUInt64 start;
-    /// End of the range (exluded).
-    NSTDUInt64 end;
-} NSTDURange;
-
 #ifdef __cplusplus
 }
 #endif