+++ include/nstd/core/str.h
@@ -2,6 +2,7 @@
 #define NSTD_CORE_STR_H_INCLUDED
 #include "char_types.h"
 #include "def.h"
+#include "range.h"
 #include "slice.h"
 #ifdef __cplusplus
 extern "C"